requirements.txt
Flask==2.1.2
line-bot-sdk==1.19.0
python-dotenv
gunicorn
